requirements.txt
torch>=2.2.0
torchvision>=0.17.0
timm>=0.9.12
tensorboard>=2.14.0
scikit-learn>=1.3.0
numpy==1.26.4
pillow>=10.0.0
opencv-python>=4.8.0
albumentations>=1.3.1
pyyaml>=6.0.1
tqdm>=4.66.0
matplotlib>=3.7.0
